BCDA-3578 further consolidated refactored code
