+++ bcda/api/requests.go
@@ -22,6 +22,7 @@ import (
 	"github.com/CMSgov/bcda-app/bcda/client"
 	"github.com/CMSgov/bcda-app/bcda/database"
 	"github.com/CMSgov/bcda-app/bcda/models"
+	"github.com/CMSgov/bcda-app/bcda/models/postgres"
 	"github.com/CMSgov/bcda-app/bcda/responseutils"
 	"github.com/CMSgov/bcda-app/bcda/servicemux"
 	"github.com/CMSgov/bcda-app/bcda/utils"
@@ -30,6 +31,19 @@ import (
 var (
 	qc *que.Client
 )
+
+func init() {
+	// Ensure that models.go is properly initialized with the service reference.
+	// As we refactor more of the code, we should be able to remove the initialization
+	// from models.go
+	cutoffDuration := time.Duration(utils.GetEnvInt("CCLF_CUTOFF_DATE_DAYS", 45)*24) * time.Hour
+	db := database.GetGORMDbConnection()
+	db.DB().SetMaxOpenConns(utils.GetEnvInt("BCDA_DB_MAX_OPEN_CONNS", 25))
+	db.DB().SetMaxIdleConns(utils.GetEnvInt("BCDA_DB_MAX_IDLE_CONNS", 25))
+	db.DB().SetConnMaxLifetime(time.Duration(utils.GetEnvInt("BCDA_DB_CONN_MAX_LIFETIME_MIN", 5)) * time.Minute)
+	repository := postgres.NewRepository(db)
+	models.GetService(repository, cutoffDuration, utils.GetEnvInt("BCDA_SUPPRESSION_LOOKBACK_DAYS", 60))
+}
 
 func BulkRequest(resourceTypes []string, w http.ResponseWriter, r *http.Request, retrieveNewBeneHistData bool) {
 	var (
@@ -289,4 +303,36 @@ func GetJobTimeout() time.Duration {
 
 func SetQC(client *que.Client) {
 	qc = client
+}
+
+// swagger:model fileItem
+type FileItem struct {
+	// FHIR resource type of file contents
+	Type string `json:"type"`
+	// URL of the file
+	URL string `json:"url"`
+}
+
+/*
+Data export job has completed successfully. The response body will contain a JSON object providing metadata about the transaction.
+swagger:response completedJobResponse
+*/
+// nolint
+type CompletedJobResponse struct {
+	// in: body
+	Body BulkResponseBody
+}
+
+type BulkResponseBody struct {
+	// Server time when the query was run
+	TransactionTime time.Time `json:"transactionTime"`
+	// URL of the bulk data export request
+	RequestURL string `json:"request"`
+	// Indicates whether an access token is required to download generated data files
+	RequiresAccessToken bool `json:"requiresAccessToken"`
+	// Information about generated data files, including URLs for downloading
+	Files []FileItem `json:"output"`
+	// Information about error files, including URLs for downloading
+	Errors []FileItem `json:"error"`
+	JobID  uint
 }
